profile page edit

+++ db/seeds.rb
@@ -1,7 +1,8 @@
 puts "Emptying database..."
 
-User.destroy_all
-Cabin.destroy_all
+User.destroy_all if Rails.env.development?
+Cabin.destroy_all if Rails.env.development?
+Booking.destroy_all if Rails.env.development?
 
 puts "Seeding database..."
